[domain] move Bucket into module
