+++ src/main/scala/net/kemitix/s3thorp/Config.scala
@@ -2,6 +2,8 @@ package net.kemitix.s3thorp
 
 import java.io.File
 
+import net.kemitix.s3thorp.domain.Bucket
+
 final case class Config(
   bucket: Bucket = Bucket(""),
   prefix: RemoteKey = RemoteKey(""),
