+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientDeleter.scala
@@ -2,7 +2,8 @@ package net.kemitix.s3thorp.awssdk
 
 import cats.effect.IO
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
-import net.kemitix.s3thorp.{Bucket, Config, DeleteS3Action, RemoteKey}
+import net.kemitix.s3thorp.domain.Bucket
+import net.kemitix.s3thorp.{Config, DeleteS3Action, RemoteKey}
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest
 
 class S3ClientDeleter(s3Client: S3CatsIOClient)
