+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientMultiPartUploader.scala
@@ -6,6 +6,7 @@ import cats.implicits._
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.{AbortMultipartUploadRequest, AmazonS3Exception, CompleteMultipartUploadRequest, CompleteMultipartUploadResult, InitiateMultipartUploadRequest, InitiateMultipartUploadResult, PartETag, UploadPartRequest, UploadPartResult}
 import net.kemitix.s3thorp._
+import net.kemitix.s3thorp.domain.Bucket
 
 import scala.util.control.NonFatal
