+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientObjectLister.scala
@@ -1,9 +1,11 @@
 package net.kemitix.s3thorp.awssdk
 
 import cats.effect.IO
-import net.kemitix.s3thorp.{Bucket, Config, HashModified, LastModified, MD5Hash, RemoteKey}
+import net.kemitix.s3thorp.{Config, HashModified, LastModified, MD5Hash, RemoteKey}
 import software.amazon.awssdk.services.s3.model.{ListObjectsV2Request, S3Object}
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
+import net.kemitix.s3thorp.domain.Bucket
+
 import scala.collection.JavaConverters._
 
 class S3ClientObjectLister(s3Client: S3CatsIOClient)
