+++ src/test/scala/net/kemitix/s3thorp/ActionGeneratorSuite.scala
@@ -2,6 +2,8 @@ package net.kemitix.s3thorp
 
 import java.time.Instant
 
+import net.kemitix.s3thorp.domain.Bucket
+
 class ActionGeneratorSuite
   extends UnitTest
     with KeyGenerator {
